refactor: simplify worker file naming by removing redundant prefixes
Renamed index-worker/ → worker/ and removed worker- prefixes from all files:
- full-indexer.js → full.js
- worker-incremental.js → incremental.js
- index-worker.js → worker.js (entry point)
- worker-admin-helpers.js → admin-helpers.js
- worker-bulk-status.js → bulk-status.js
- worker-fetch.js → fetch.js
- worker-linked-content.js → linked-content.js
- worker-parse.js → parse.js

Updated all imports within worker files and load.js to reference new paths.
The indexing/ directory context already implies worker execution, making both the directory name and file prefixes redundant.
