+++ nx/blocks/media-library/indexing/worker/fetch.js
